+++ src/objects/CauseOfIsState.java
@@ -3,7 +3,7 @@ package objects;
 public class CauseOfIsState extends Annotation {
 	private String causeState, effectEvent;
 
-	public CauseOfIsState(int startNode, String causeState, String effectEvent) {
+	public CauseOfIsState(long startNode, String causeState, String effectEvent) {
 		super(startNode);
 		this.causeState = causeState;
 		this.effectEvent = effectEvent;
