+++ src/objects/HappensRelation.java
@@ -3,7 +3,7 @@ package objects;
 public class HappensRelation extends Annotation {
 	private String timeHappened;
 
-	public HappensRelation(int startNode, String timeHappened) {
+	public HappensRelation(long startNode, String timeHappened) {
 		super(startNode);
 		this.timeHappened = timeHappened;
 	}
